[Localization] Extract testing fixture and diagnostic definitions into a header
Doing so make it easy to share common bits of localization testing
infrastructure.
