+++ unittests/Localization/DefToYAMLConverterTests.cpp
@@ -10,6 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "LocalizationTest.h"
 #include "swift/Localization/LocalizationFormat.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/SmallString.h"
@@ -19,8 +20,6 @@
 #include "llvm/Support/Path.h"
 #include "llvm/Support/Signals.h"
 #include "llvm/Support/ToolOutputFile.h"
-#include "llvm/Support/YAMLParser.h"
-#include "llvm/Support/YAMLTraits.h"
 #include "llvm/Support/raw_ostream.h"
 #include "gtest/gtest.h"
 #include <cstdlib>
@@ -30,22 +29,7 @@
 
 using namespace swift;
 using namespace swift::diag;
-
-enum LocalDiagID : uint32_t {
-#define DIAG(KIND, ID, Options, Text, Signature) ID,
-#include "swift/AST/DiagnosticsAll.def"
-  NumDiags
-};
-
-static constexpr const char *const diagnosticID[] = {
-#define DIAG(KIND, ID, Options, Text, Signature) #ID,
-#include "swift/AST/DiagnosticsAll.def"
-};
-
-static constexpr const char *const diagnosticMessages[] = {
-#define DIAG(KIND, ID, Options, Text, Signature) Text,
-#include "swift/AST/DiagnosticsAll.def"
-};
+using namespace swift::unittests;
 
 static std::string getMainExecutablePath() {
   llvm::StringRef libPath = llvm::sys::path::parent_path(SWIFTLIB_DIR);
@@ -62,50 +46,7 @@ static std::string getDefaultLocalizationPath() {
   return std::string(DefaultDiagnosticMessagesDir);
 }
 
-struct DefToYAMLConverterTest : public ::testing::Test {
-  std::string YAMLPath;
-
-public:
-  DefToYAMLConverterTest() {
-    llvm::SmallString<128> tempFilename;
-    std::error_code error =
-        llvm::sys::fs::createTemporaryFile("en", "yaml", tempFilename);
-    assert(!error);
-
-    YAMLPath = std::string(tempFilename);
-  }
-
-  void SetUp() override {
-    bool failed = convertDefIntoYAML(YAMLPath);
-    assert(!failed && "failed to generate a YAML file");
-  }
-
-  void TearDown() override { llvm::sys::fs::remove(YAMLPath); }
-
-  /// Random number in [0,n)
-  unsigned RandNumber(unsigned n) { return unsigned(rand()) % n; }
-
-protected:
-  static bool convertDefIntoYAML(std::string outputPath) {
-    std::error_code error;
-    llvm::raw_fd_ostream OS(outputPath, error, llvm::sys::fs::F_None);
-    if (OS.has_error() || error)
-      return true;
-
-    llvm::ArrayRef<const char *> ids(diagnosticID, LocalDiagID::NumDiags);
-    llvm::ArrayRef<const char *> messages(diagnosticMessages,
-                                          LocalDiagID::NumDiags);
-
-    DefToYAMLConverter converter(ids, messages);
-    converter.convert(OS);
-
-    OS.flush();
-
-    return OS.has_error();
-  }
-};
-
-TEST_F(DefToYAMLConverterTest, MissingLocalizationFiles) {
+TEST_F(LocalizationTest, MissingLocalizationFiles) {
   ASSERT_TRUE(llvm::sys::fs::exists(getDefaultLocalizationPath()));
   llvm::SmallString<128> EnglishLocalization(getDefaultLocalizationPath());
   llvm::sys::path::append(EnglishLocalization, "en");
@@ -115,7 +56,7 @@ TEST_F(DefToYAMLConverterTest, MissingLocalizationFiles) {
   ASSERT_TRUE(llvm::sys::fs::exists(EnglishLocalization));
 }
 
-TEST_F(DefToYAMLConverterTest, MatchDiagnosticMessagesSequentially) {
+TEST_F(LocalizationTest, ConverterTestMatchDiagnosticMessagesSequentially) {
   YAMLLocalizationProducer yaml(YAMLPath);
   yaml.forEachAvailable([](swift::DiagID id, llvm::StringRef translation) {
     llvm::StringRef msg = diagnosticMessages[static_cast<uint32_t>(id)];
@@ -123,7 +64,7 @@ TEST_F(DefToYAMLConverterTest, MatchDiagnosticMessagesSequentially) {
   });
 }
 
-TEST_F(DefToYAMLConverterTest, MatchDiagnosticMessagesRandomly) {
+TEST_F(LocalizationTest, ConverterTestMatchDiagnosticMessagesRandomly) {
   YAMLLocalizationProducer yaml(YAMLPath);
 
   std::random_device rd;
